refactor: rename to `console-settings.component ` to respect angular convention
