+++ src/organization/configuration/console/console-settings.component.scss
