+++ src/organization/configuration/console/console-settings.component.spec.ts
@@ -28,7 +28,7 @@ import { HttpTestingController } from '@angular/common/http/testing';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { InteractivityChecker } from '@angular/cdk/a11y';
 
-import { ConsoleSettingsComponent } from './console-settings';
+import { ConsoleSettingsComponent } from './console-settings.component';
 
 import { CONSTANTS_TESTING, GioHttpTestingModule } from '../../../shared/testing';
 import { OrganizationSettingsModule } from '../organization-settings.module';
