+++ src/organization/configuration/organization-settings.module.ts
@@ -30,7 +30,7 @@ import { MatTooltipModule } from '@angular/material/tooltip';
 import { MatDialogModule } from '@angular/material/dialog';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 
-import { ConsoleSettingsComponent } from './console/console-settings';
+import { ConsoleSettingsComponent } from './console/console-settings.component';
 
 import { GioConfirmDialogModule } from '../../shared/components/confirm-dialog/confirm-dialog.module';
 @NgModule({
